Popin - Splitting overlay component
