+++ src/components/popin/index.js
@@ -1,58 +1,10 @@
 // Dependencies
-import React, { PropTypes } from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom';
 import builder from 'focus-core/component/builder';
 import types from 'focus-core/component/types';
 import { ArgumentInvalidException } from 'focus-core/exception';
-
-/**
-* Small overlay component used to listen to scroll and prevent it to leave the Popin component
-*/
-const Overlay = React.createClass({
-    displayName: 'PopinOverlay',
-    propTypes: {
-        children: PropTypes.object,
-        clickHandler: PropTypes.func,
-        show: PropTypes.bool
-    },
-    getDefaultProps() {
-        return { show: false };
-    },
-    /**
-    * Store the body overgflow property, and set it to hidden
-    * @private
-    */
-    _hideBodyOverflow() {
-        document.body.style['overflow-y'] = 'hidden';
-    },
-    /**
-    * Restore body overflow property
-    * @private
-    */
-    _restoreBodyOverflow() {
-        document.body.style['overflow-y'] = 'visible';
-    },
-    /**
-    * Component will unmount event handler.
-    * Remove the mouse wheel listener.
-    */
-    componentWillUnmount() {
-        // ReactDOM.findDOMNode(this.refs.overlay).removeEventListener('mousewheel', this._onScroll);
-        this._restoreBodyOverflow();
-    },
-    /**
-    * Render the component
-    * @return {XML} the rendered HTML
-    */
-    render() {
-        const { children, clickHandler, show } = this.props;
-        return (
-            <div className='animated fadeIn' data-animation='fadeIn' data-closing-animation='fadeOut' data-focus='popin-overlay' data-visible={show} onClick={clickHandler} ref='overlay'>
-                {children}
-            </div>
-        );
-    }
-});
+import Overlay from './overlay';
 
 /**
 * The popin component configuration
